style(backend): Move logger to modules
