+++ backend/modules/logger/index.js
@@ -1,6 +1,8 @@
 /**
  * File for defining logger
  * so you can use it throughout your project
+ *
+ * See more info about logger class at https://github.com/pinojs/pino
  */
 const pino = require('pino');
